Added react navigation to app
