+++ src/App.tsx
@@ -1,7 +1,19 @@
-import { Font  } from 'expo';
+import { Font, Constants  } from 'expo';
 import { Container, Content, Header } from 'native-base';
 import * as React from 'react';
-import HomeScreen from './pages/Home';
+import HomeScreen from './pages/home/index';
+import { View, StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  statusBar: {
+    backgroundColor: '#C2185B',
+    height: Constants.statusBarHeight,
+  },
+  mainView: {
+    flex: 1
+  }
+  // rest of the styles
+});
 
 export default class App extends React.Component <any, any> {
   constructor(props: {}) {
@@ -15,6 +27,7 @@ export default class App extends React.Component <any, any> {
     await Font.loadAsync({
       Roboto: require('native-base/Fonts/Roboto.ttf'),
       Roboto_medium: require('native-base/Fonts/Roboto_medium.ttf'),
+      Ionicons: require('native-base/Fonts/Ionicons.ttf')
     });
 
     this.setState(Object.assign(this.state, { isReady: true }));
@@ -26,7 +39,10 @@ export default class App extends React.Component <any, any> {
     }
 
     return (
-      <HomeScreen />
+      <View style={styles.mainView}>
+        <View style={styles.statusBar} />
+      <HomeScreen/>
+      </View>
     );
   }
 }
