--- src/pages/Home.tsx
@@ -1,51 +1,0 @@
-import * as React from 'react';
-import { StatusBar } from 'react-native';
-import {
-  Container,
-  Header,
-  Title,
-  Left,
-  Icon,
-  Right,
-  Button,
-  Body,
-  Content,
-  Text,
-  Card,
-  CardItem } from 'native-base';
-
-export default class HomeScreen extends React.Component<any, any> {
-  public render() {
-    return (
-      <Container>
-        <Header>
-          <Left>
-            <Button
-              transparent
-              onPress={() => this.props.navigation.navigate('DrawerOpen')}>
-              <Icon name="menu" />
-            </Button>
-          </Left>
-          <Body>
-            <Title>HomeScreen</Title>
-          </Body>
-          <Right />
-        </Header>
-        <Content padder>
-          <Card>
-            <CardItem>
-              <Body>
-                <Text>Chat App to talk some awesome people!</Text>
-              </Body>
-            </CardItem>
-          </Card>
-          <Button full rounded dark
-            style={{ marginTop: 10 }}
-            onPress={() => this.props.navigation.navigate('VancomycinEmpiricDosing')}>
-            <Text>Vancomycin Empiric Dosing</Text>
-          </Button>
-        </Content>
-      </Container>
-    );
-  }
-}
